mx28evk: Fixed compiler warnings and section mismatches
Fixed warnings reported by the compiler and section mismatches reported
by the linker.  This was done as part of the effort to port the mx28evk
from the 2.6.31 kernel to the 2.6.35 kernel.

+++ arch/arm/mm/init.c
@@ -629,9 +629,9 @@ void __init mem_init(void)
 				(PAGE_SIZE)),
 			MLK(FIXADDR_START, FIXADDR_TOP),
 #ifdef CONFIG_MMU
-			MLM(CONSISTENT_BASE, CONSISTENT_END),
+			MLM((unsigned long) CONSISTENT_BASE, (unsigned long) CONSISTENT_END),
 #endif
-			MLM(VMALLOC_START, VMALLOC_END),
+			MLM((unsigned long) VMALLOC_START, (unsigned long) VMALLOC_END),
 			MLM(PAGE_OFFSET, (unsigned long)high_memory),
 #ifdef CONFIG_HIGHMEM
 			MLM(PKMAP_BASE, (PKMAP_BASE) + (LAST_PKMAP) *
